모든 시간을 int로 바꿈
부동소숫점 오차가 우려되어서 int로 바꿈
근데 그러면 1000fps를 넘었을때 dt가 0이됨! 이거 어캄

+++ demo/benchmark.cpp
@@ -70,13 +70,13 @@ int main() {
     // 현재 테스트에 생성된 Object들의 ID
     std::vector<unsigned int> objectIds;
 
-    constexpr float WARMUP_TIME = 2000.0f;
-    constexpr float TEST_TIME = 5000.0f;
+    constexpr int WARMUP_TIME = 2000;
+    constexpr int TEST_TIME = 5000;
 
     int benchmarkIndex = 0;
 
-    float elapsedTime = 0.0f;
-    float measuredTime = 0.0f;
+    int elapsedTime = 0;
+    int measuredTime = 0;
 
     uint64_t frameCount = 0;
 
@@ -162,8 +162,8 @@ int main() {
 
         createObjects(objectCount);
 
-        elapsedTime = 0.0f;
-        measuredTime = 0.0f;
+        elapsedTime = 0;
+        measuredTime = 0;
         frameCount = 0;
 
         measuring = false;
@@ -191,7 +191,7 @@ int main() {
         }
     });
 
-    scene.setUpdateCallback([&](knot::Scene& currentScene, float deltaTime) {
+    scene.setUpdateCallback([&](knot::Scene& currentScene, int deltaTime) {
         if (finished)
             return;
 
@@ -207,8 +207,8 @@ int main() {
             if (elapsedTime >= WARMUP_TIME) {
                 measuring = true;
 
-                elapsedTime = 0.0f;
-                measuredTime = 0.0f;
+                elapsedTime = 0;
+                measuredTime = 0;
                 frameCount = 0;
 
                 std::cout << "Measuring...\n";
@@ -252,7 +252,7 @@ int main() {
 
         // TEST_TIME 동안 2 * M_PI 라디안을 도는 각속도 계산
         constexpr float TWO_PI = 6.28318530717958647692f;
-        const float angularVelocity = TWO_PI / TEST_TIME;
+        const float angularVelocity = TWO_PI / static_cast<float>(TEST_TIME);
         const float angle = t * angularVelocity;
 
         camera.position.x = std::sin(angle) * 20.0f;
